psutil_extra: Check for bad resource values on FreeBSD and NetBSD
This makes the errors raised more consistent across platforms.

The Linux and DragonflyBSD implementations already check for this case
(on Linux we delegate to resource.prlimit() to do the check; on
DragonflyBSD it's done manually because we actually need to do manual
checking of the resource values on that platform).

However, on FreeBSD and NetBSD, this case currently results in an
OSError being raised, which doesn't match the behavior of the other
platforms (which raise a ValueError).

Now, all platforms raise a ValueError("invalid resource specified") in
this case.