_psfreebsd: Handle operations on PID 0 properly
On FreeBSD, it seems that the kernel shows up as PID 0, and some
operations can actually be performed on it. This commit makes the
functions that can be implemented on PID 0 work properly.

OpenBSD doesn't seem to do the same, and I haven't been able to test on
the other BSDs, so for now this is just FreeBSD.